requirements.txt
tensorflow
numpy
matplotlib
seaborn
google-generativeai
cbor2
python-dotenv
